backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pydantic==2.11.7
typing==3.7.4.3
requests==2.32.5
aiohttp==3.12.15

SQLAlchemy==1.4.11
psycopg2==2.8.6
python-dotenv==1.0.0